chore: move useGenerate hook to proper place, update references and imports
